完善更新书本信息

+++ src/main/kotlin/com/setruth/domain/AdminInfo.kt
@@ -1,5 +1,5 @@
 package com.setruth.domain
 
-class AdminInfo(var id: Int,  var account: String,  var password: String, var lastLoginTime: Int) {
+class AdminInfo(var id: Int,  var account: String,  var password: String, var lastLoginTime: Long) {
 
 }
